codice/PythonRaspberryBridge/requirements.txt
paho-mqtt
getmac
requests
opencv-python
opencv-contrib-python
urllib3
tcping
numpy